Rename ValueControWidgetConnection to ControlParameterWidgetConnection and PropertyControlWidgetConnection to ControlWidgetPropertyConnection.
Also, move ControlWidgetConnection and friends out of wbasewidget.cpp/h.
